+++ presentation/src/main/java/com/sms/presentation/main/ui/login/LoginActivity.kt
@@ -1,0 +1,70 @@
+package com.sms.presentation.main.ui.login
+
+import android.content.Intent
+import android.os.Bundle
+import android.util.DisplayMetrics
+import android.util.Log
+import android.view.Display
+import androidx.activity.ComponentActivity
+import androidx.activity.compose.setContent
+import androidx.activity.viewModels
+import androidx.compose.foundation.layout.Box
+import com.msg.gauthsignin.GAuthSigninWebView
+import com.sms.presentation.BuildConfig
+import com.sms.presentation.main.ui.MainActivity
+import com.sms.presentation.main.ui.login.component.LoginButton
+import com.sms.presentation.main.ui.login.component.LoginPageBackGround
+import com.sms.presentation.main.ui.login.component.TopComponent
+import com.sms.presentation.main.viewmodel.LoginViewModel
+import com.sms.presentation.main.viewmodel.util.Event
+import dagger.hilt.android.AndroidEntryPoint
+
+@AndroidEntryPoint
+class LoginActivity : ComponentActivity() {
+    private val viewModel by viewModels<LoginViewModel>()
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        observeEvent()
+        setContent {
+            val display: Display = windowManager.defaultDisplay
+            val outMetrics = DisplayMetrics()
+            display.getMetrics(outMetrics)
+            val density = resources.displayMetrics.density
+            val dpWidth = outMetrics.widthPixels / density
+
+            Box {
+                LoginPageBackGround()
+                TopComponent()
+                LoginButton(dpWidth = dpWidth) {
+                    setContent {
+                        GAuthSigninWebView(
+                            clientId = BuildConfig.CLIENT_ID,
+                            redirectUri = BuildConfig.REDIRECT_URI
+                        ) { code ->
+                            viewModel.gAuthLogin(code = code)
+                        }
+                    }
+                }
+            }
+        }
+    }
+
+    private fun observeEvent() {
+        observeLoginEvent()
+    }
+
+    private fun observeLoginEvent() {
+        viewModel.gAuthLoginRequest.observe(this) { event ->
+            when (event) {
+                Event.Success -> {
+                    startActivity(Intent(this, MainActivity::class.java))
+                    finish()
+                }
+                else -> {
+                    Log.d("login", event.toString())
+                }
+            }
+        }
+    }
+}
